Actually Ed: swapped the order of body upgrades
Upgraded Legs now comes first, since it allows farming for Ka.

It also recommends liver-before-stomach if the player picked
the astral booze, though perhaps this should always be the
case, since it allows for overdrinking.

Armor plating before Even More Elemental Wards, since the
latter is primarily useful for A-boo, but armor plating
could be the difference between an extra turn flyering
in the junkyard.